LayoutTests/accessibility/win/list-role-expected.txt

This test verifies that, on Windows, lists return the "list" role.

Bug 32736 - REGRESSION(r52233): MSAA: Accessibility role of lists is wrong

list item
PASS list.role is "AXRole: AXList"
